+++ qualtran/bloqs/arithmetic/__init__.py
@@ -23,7 +23,11 @@ from qualtran.bloqs.arithmetic.comparison import (
     LessThanEqual,
     SingleQubitCompare,
 )
-from qualtran.bloqs.arithmetic.conversions import SignedIntegerToTwosComplement, ToContiguousIndex
+from qualtran.bloqs.arithmetic.conversions import (
+    SignedIntegerToTwosComplement,
+    SignExtend,
+    ToContiguousIndex,
+)
 from qualtran.bloqs.arithmetic.hamming_weight import HammingWeightCompute
 from qualtran.bloqs.arithmetic.multiplication import (
     MultiplyTwoReals,
@@ -35,7 +39,6 @@ from qualtran.bloqs.arithmetic.multiplication import (
     SumOfSquares,
 )
 from qualtran.bloqs.arithmetic.negate import Negate
-from qualtran.bloqs.arithmetic.sign_extension import SignExtend
 from qualtran.bloqs.arithmetic.sorting import BitonicSort, Comparator
 from qualtran.bloqs.arithmetic.subtraction import Subtract
