+++ qualtran/bloqs/arithmetic/conversions/sign_extension_test.py
@@ -11,11 +11,27 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#      https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
 import numpy as np
 import pytest
 
 from qualtran import BloqBuilder, QInt, QUInt
-from qualtran.bloqs.arithmetic.sign_extension import _sign_extend, _sign_extend_fxp, SignExtend
+from qualtran.bloqs.arithmetic.conversions.sign_extension import (
+    _sign_extend,
+    _sign_extend_fxp,
+    SignExtend,
+)
 from qualtran.bloqs.basic_gates import IntEffect, IntState
